Fix groq/LLama profile
As the creator of the Groq PR, and the groq.js file for the bots, I explicitly mentioned that my Groq implementation never had support for embeddings. There is a reason that was there. To my knowledge, and my memory of the Groq documentation, it never had support for embeddings like this. Please don't do this again :(